refactor: decomposition selection utils
