+++ packages/store/src/public_api.ts
@@ -29,7 +29,7 @@ export {
 } from './symbols';
 export { Selector } from './decorators/selector';
 export { getActionTypeFromInstance, actionMatcher } from './utils/utils';
-export { createSelector } from './utils/selector-utils';
+export { createSelector } from './utils/selector/create-selector';
 export { NgxsExecutionStrategy } from './execution/symbols';
 export { ActionType, ActionOptions } from './actions/symbols';
 export { NoopNgxsExecutionStrategy } from './execution/noop-ngxs-execution-strategy';
